backend/requirements.txt
fastapi[all]>=0.104.1
uvicorn[standard]>=0.24.0
motor>=3.3.2
pymongo>=4.6.0
python-multipart>=0.0.6
qdrant-client>=1.15.1
sentence-transformers>=2.2.2
transformers>=4.36.2
torch>=2.6.0
accelerate>=0.25.0
redis>=5.0.1
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
pydantic>=2.5.2
aiofiles>=23.2.1
websockets>=12.0
beautifulsoup4>=4.12.2
PyPDF2>=3.0.1
python-docx>=1.1.0
pillow>=10.1.0
spacy>=3.7.2
tiktoken>=0.5.2
pandas>=2.1.4
numpy>=1.26.2
scikit-learn>=1.3.2
httpx>=0.26.0
rank-bm25>=0.2.2
aioredis>=2.0.1
passlib[bcrypt]>=1.7.4
pytest>=7.4.3
pytest-asyncio>=0.21.1
loguru>=0.7.0
